feat(clientcmd,faro): extract client commands into reusable clientcmd package and add faro slim client
Move authentication, context, playbook, plugin, connection, and whoami commands from cmd package into a new clientcmd package to enable code reuse across binaries. This allows the slim faro client to leverage the same remote-only command surfaces without duplicating code.

Add faro binary: a lightweight Mission Control client that operates exclusively against a remote server via the SDK, with no embedded UI or database. Faro supports catalog browsing, playbook execution, plugin operations, and connection management through the remote API.

Introduce LocalConnectionOps, LocalRunHandler, LocalPluginDispatch, and LocalPluginRefresh interfaces to allow the full mission-control binary to inject local-only functionality (database operations, local playbook execution, local plugin dispatch) while keeping clientcmd generic.

Update Makefile and Taskfile to build faro binaries for linux amd64/arm64, darwin arm64, and windows amd64. Include faro artifacts in release builds.

BREAKING CHANGE: cmd package functions are now in clientcmd package with capitalized names (e.g., newAPIClient → NewAPIClient). Full binary imports must update references and wire LocalConnectionOps, LocalRunHandler, and LocalPluginDispatch.

Refs #faro
