+++ clientcmd/connection_browser_teardown.go
@@ -1,4 +1,4 @@
-package cmd
+package clientcmd
 
 import (
 	gocontext "context"
