+++ clientcmd/connection_browser_unix.go
@@ -1,6 +1,6 @@
 //go:build !windows
 
-package cmd
+package clientcmd
 
 import (
 	"os"
