+++ clientcmd/connection_flags.go
@@ -1,470 +1,13 @@
-package cmd
+package clientcmd
 
 import (
-	gocontext "context"
-	"errors"
 	"fmt"
-	"os"
-	"path/filepath"
 
-	"github.com/aws/aws-sdk-go-v2/config"
-	"github.com/flanksource/clicky"
-	"github.com/flanksource/duty"
 	"github.com/flanksource/duty/models"
-	"github.com/flanksource/duty/shutdown"
-	"github.com/google/uuid"
 	"github.com/spf13/cobra"
-	"gorm.io/gorm"
-
-	"github.com/flanksource/incident-commander/connection"
 )
 
-var Connection = &cobra.Command{
-	Use:          "connection",
-	Short:        "Manage connections",
-	SilenceUsage: true,
-}
-
-var ConnectionAdd = &cobra.Command{
-	Use:   "add",
-	Short: "Add a new connection",
-	Long: `Add a new connection to the database.
-
-Examples:
-  app connection add slack --name test --namespace default --token xoxb-xxx --channel C12345
-  app connection add postgres --name mydb --namespace default --url "postgres://user:pass@localhost:5432/db"
-  app connection add aws --name my-aws --namespace default --access-key AKIA... --secret-key ... --region us-east-1
-  app connection add postgres --name mydb --namespace default --url "..." --test`,
-	PersistentPreRun:  PreRun,
-	SilenceUsage:      true,
-	DisableAutoGenTag: true,
-}
-
-type connectionTypeSpec struct {
-	Name    string
-	Type    string
-	Aliases []string
-	Short   string
-}
-
-var connectionAddTypeSpecs = []connectionTypeSpec{
-	{Name: "slack", Type: models.ConnectionTypeSlack, Short: "Add a Slack connection"},
-	{Name: "postgres", Type: models.ConnectionTypePostgres, Short: "Add a Postgres connection"},
-	{Name: "mysql", Type: models.ConnectionTypeMySQL, Short: "Add a MySQL connection"},
-	{Name: "mssql", Type: models.ConnectionTypeSQLServer, Aliases: []string{"sqlserver", "sql-server"}, Short: "Add a SQL Server connection"},
-	{Name: "mongo", Type: models.ConnectionTypeMongo, Aliases: []string{"mongodb"}, Short: "Add a MongoDB connection"},
-	{Name: "aws", Type: models.ConnectionTypeAWS, Short: "Add an AWS connection"},
-	{Name: "aws-kms", Type: models.ConnectionTypeAWSKMS, Aliases: []string{"awskms"}, Short: "Add an AWS KMS connection"},
-	{Name: "s3", Type: models.ConnectionTypeS3, Short: "Add an S3 connection"},
-	{Name: "azure", Type: models.ConnectionTypeAzure, Short: "Add an Azure connection"},
-	{Name: "azure-keyvault", Type: models.ConnectionTypeAzureKeyVault, Aliases: []string{"azure-key-vault"}, Short: "Add an Azure Key Vault connection"},
-	{Name: "azure-devops", Type: models.ConnectionTypeAzureDevops, Aliases: []string{"azuredevops"}, Short: "Add an Azure DevOps connection"},
-	{Name: "gcp", Type: models.ConnectionTypeGCP, Short: "Add a GCP connection"},
-	{Name: "gcs", Type: models.ConnectionTypeGCS, Short: "Add a GCS connection"},
-	{Name: "gcp-kms", Type: models.ConnectionTypeGCPKMS, Aliases: []string{"gcpkms"}, Short: "Add a GCP KMS connection"},
-	{Name: "facet", Type: models.ConnectionTypeFacet, Short: "Add a Facet connection"},
-	{Name: "discord", Type: models.ConnectionTypeDiscord, Short: "Add a Discord connection"},
-	{Name: "smtp", Type: models.ConnectionTypeEmail, Aliases: []string{"email"}, Short: "Add an SMTP connection"},
-	{Name: "telegram", Type: models.ConnectionTypeTelegram, Short: "Add a Telegram connection"},
-	{Name: "ntfy", Type: models.ConnectionTypeNtfy, Short: "Add an Ntfy connection"},
-	{Name: "pushbullet", Type: models.ConnectionTypePushbullet, Short: "Add a Pushbullet connection"},
-	{Name: "pushover", Type: models.ConnectionTypePushover, Short: "Add a Pushover connection"},
-	{Name: "http", Type: models.ConnectionTypeHTTP, Short: "Add an HTTP connection"},
-	{Name: "git", Type: models.ConnectionTypeGit, Short: "Add a Git connection"},
-	{Name: "github", Type: models.ConnectionTypeGithub, Short: "Add a GitHub connection"},
-	{Name: "gitlab", Type: models.ConnectionTypeGitlab, Short: "Add a GitLab connection"},
-	{Name: "kubernetes", Type: models.ConnectionTypeKubernetes, Short: "Add a Kubernetes connection"},
-	{Name: "folder", Type: models.ConnectionTypeFolder, Short: "Add a folder connection"},
-	{Name: "sftp", Type: models.ConnectionTypeSFTP, Short: "Add an SFTP connection"},
-	{Name: "smb", Type: models.ConnectionTypeSMB, Short: "Add an SMB connection"},
-	{Name: "prometheus", Type: models.ConnectionTypePrometheus, Short: "Add a Prometheus connection"},
-	{Name: "loki", Type: models.ConnectionTypeLoki, Short: "Add a Loki connection"},
-	{Name: "openai", Type: models.ConnectionTypeOpenAI, Short: "Add an OpenAI connection"},
-	{Name: "anthropic", Type: models.ConnectionTypeAnthropic, Short: "Add an Anthropic connection"},
-	{Name: "ollama", Type: models.ConnectionTypeOllama, Short: "Add an Ollama connection"},
-	{Name: "gemini", Type: models.ConnectionTypeGemini, Short: "Add a Gemini connection"},
-}
-
-type connectionFlags struct {
-	Name      string
-	Namespace string
-	Type      string
-	Test      bool
-	DryRun    bool
-
-	// Common fields
-	URL         string
-	Username    string
-	Password    string
-	Certificate string
-	InsecureTLS bool
-
-	// Kubernetes
-	Kubeconfig string
-
-	// AWS
-	AccessKey    string
-	SecretKey    string
-	Region       string
-	Profile      string
-	FromProfile  string
-	SessionToken string
-
-	// AWS KMS / GCP KMS / Azure Key Vault
-	KeyID string
-
-	// S3 / GCS
-	Bucket       string
-	UsePathStyle bool
-
-	// Azure
-	ClientID     string
-	ClientSecret string
-	TenantID     string
-
-	// Azure DevOps
-	Organization        string
-	PersonalAccessToken string
-
-	// Git
-	Ref string
-
-	// Slack
-	Channel  string
-	Token    string
-	BotName  string
-	Color    string
-	Icon     string
-	ThreadTS string
-	Title    string
-
-	// SMTP
-	Host        string
-	Port        int
-	FromAddress string
-	FromName    string
-	ToAddresses []string
-	Subject     string
-	Auth        string
-	Encryption  string
-
-	// Telegram / Ntfy
-	Chats string
-	Topic string
-
-	// Pushbullet / Pushover
-	Targets []string
-	User    string
-
-	// Discord
-	WebhookID string
-
-	// SFTP / SMB
-	Path  string
-	Share string
-
-	// Mongo
-	ReplicaSet string
-
-	// MSSQL
-	TrustServerCertificate bool
-
-	// Database common
-	Database string
-
-	// HTTP / Prometheus auth
-	Bearer            string
-	OAuthClientID     string
-	OAuthClientSecret string
-	OAuthTokenURL     string
-	OAuthScopes       string
-	TLSCA             string
-	TLSCert           string
-	TLSKey            string
-
-	// Facet
-	TimestampURL string
-
-	// AI models
-	Model  string
-	ApiKey string
-
-	// Properties (generic key=value pairs)
-	Properties []string
-}
-
-func kubeconfigDefault() string {
-	if v := os.Getenv("KUBECONFIG"); v != "" {
-		return v
-	}
-	if home, err := os.UserHomeDir(); err == nil {
-		return filepath.Join(home, ".kube", "config")
-	}
-	return ""
-}
-
-func validateConnectionFlags(flags *connectionFlags) error {
-	switch flags.Type {
-	case models.ConnectionTypeSlack:
-		if flags.Token == "" {
-			return fmt.Errorf("--token is required for Slack connections")
-		}
-		if flags.Channel == "" {
-			return fmt.Errorf("--channel is required for Slack connections")
-		}
-
-	case models.ConnectionTypePostgres, models.ConnectionTypeMySQL, models.ConnectionTypeSQLServer, models.ConnectionTypeMongo:
-		if flags.URL == "" && flags.Host == "" {
-			return fmt.Errorf("--url or --host is required for database connections")
-		}
-
-	case models.ConnectionTypeAWS, models.ConnectionTypeAWSKMS, models.ConnectionTypeS3:
-		// AWS can use instance profiles, so credentials are optional
-
-	case models.ConnectionTypeAzure, models.ConnectionTypeAzureKeyVault:
-		if flags.ClientID == "" || flags.ClientSecret == "" || flags.TenantID == "" {
-			return fmt.Errorf("--client-id, --client-secret, and --tenant-id are required for Azure connections")
-		}
-
-	case models.ConnectionTypeAzureDevops:
-		if flags.PersonalAccessToken == "" {
-			return fmt.Errorf("--personal-access-token is required for Azure DevOps connections")
-		}
-
-	case models.ConnectionTypeGCP, models.ConnectionTypeGCS, models.ConnectionTypeGCPKMS:
-		// GCP can use default credentials, so certificate is optional
-
-	case models.ConnectionTypeDiscord:
-		if flags.WebhookID == "" || flags.Token == "" {
-			return fmt.Errorf("--webhook-id and --token are required for Discord connections")
-		}
-
-	case models.ConnectionTypeEmail:
-		if flags.Host == "" {
-			return fmt.Errorf("--host is required for SMTP connections")
-		}
-
-	case models.ConnectionTypeTelegram:
-		if flags.Token == "" {
-			return fmt.Errorf("--token is required for Telegram connections")
-		}
-
-	case models.ConnectionTypeNtfy:
-		if flags.Host == "" || flags.Topic == "" {
-			return fmt.Errorf("--host and --topic are required for Ntfy connections")
-		}
-
-	case models.ConnectionTypePushbullet:
-		if flags.Token == "" {
-			return fmt.Errorf("--token is required for Pushbullet connections")
-		}
-
-	case models.ConnectionTypePushover:
-		if flags.Token == "" || flags.User == "" {
-			return fmt.Errorf("--token and --user are required for Pushover connections")
-		}
-
-	case models.ConnectionTypeHTTP:
-		if flags.URL == "" {
-			return fmt.Errorf("--url is required for HTTP connections")
-		}
-
-	case models.ConnectionTypeGit:
-		if flags.URL == "" {
-			return fmt.Errorf("--url is required for Git connections")
-		}
-
-	case models.ConnectionTypeGithub, models.ConnectionTypeGitlab:
-		if flags.PersonalAccessToken == "" {
-			return fmt.Errorf("--personal-access-token is required for GitHub/GitLab connections")
-		}
-
-	case models.ConnectionTypeFolder:
-		if flags.Path == "" {
-			return fmt.Errorf("--path is required for folder connections")
-		}
-
-	case models.ConnectionTypeSFTP:
-		if flags.Host == "" {
-			return fmt.Errorf("--host is required for SFTP connections")
-		}
-
-	case models.ConnectionTypeSMB:
-		if flags.Host == "" || flags.Share == "" {
-			return fmt.Errorf("--host and --share are required for SMB connections")
-		}
-
-	case models.ConnectionTypePrometheus, models.ConnectionTypeLoki:
-		if flags.URL == "" {
-			return fmt.Errorf("--url is required for %s connections", flags.Type)
-		}
-
-	case models.ConnectionTypeOpenAI, models.ConnectionTypeAnthropic, models.ConnectionTypeGemini:
-		if flags.ApiKey == "" {
-			return fmt.Errorf("--api-key is required for %s connections", flags.Type)
-		}
-
-	case models.ConnectionTypeOllama:
-		if flags.URL == "" {
-			return fmt.Errorf("--url is required for Ollama connections")
-		}
-
-	case models.ConnectionTypeFacet:
-		if flags.URL == "" {
-			return fmt.Errorf("--url is required for Facet connections")
-		}
-	}
-
-	return nil
-}
-
-func loadAWSProfile(flags *connectionFlags) error {
-	cfg, err := config.LoadDefaultConfig(gocontext.Background(),
-		config.WithSharedConfigProfile(flags.FromProfile),
-	)
-	if err != nil {
-		return fmt.Errorf("failed to load AWS profile %q: %w", flags.FromProfile, err)
-	}
-
-	creds, err := cfg.Credentials.Retrieve(gocontext.Background())
-	if err != nil {
-		return fmt.Errorf("failed to retrieve credentials from AWS profile %q: %w", flags.FromProfile, err)
-	}
-
-	if flags.AccessKey == "" {
-		flags.AccessKey = creds.AccessKeyID
-	}
-	if flags.SecretKey == "" {
-		flags.SecretKey = creds.SecretAccessKey
-	}
-	if creds.SessionToken != "" && flags.SessionToken == "" {
-		flags.SessionToken = creds.SessionToken
-	}
-	if flags.Region == "" && cfg.Region != "" {
-		flags.Region = cfg.Region
-	}
-	return nil
-}
-
-func runConnectionAdd(flags *connectionFlags) error {
-	if flags.FromProfile != "" {
-		if err := loadAWSProfile(flags); err != nil {
-			return err
-		}
-	}
-
-	if flags.DryRun {
-		out, err := marshalDryRunOutput(flags)
-		if err != nil {
-			return fmt.Errorf("failed to marshal dry-run output: %w", err)
-		}
-		fmt.Print(string(out))
-		return nil
-	}
-
-	if err := validateConnectionFlags(flags); err != nil {
-		return err
-	}
-
-	conn, err := buildConnectionFromFlags(flags)
-	if err != nil {
-		return fmt.Errorf("failed to build connection: %w", err)
-	}
-
-	if mcCtx, ok := contextHasAPI(); ok {
-		return runConnectionAddViaAPI(mcCtx, flags, &conn)
-	}
-	return runConnectionAddViaDB(flags, &conn)
-}
-
-func runConnectionAddViaAPI(mcCtx *MCContext, flags *connectionFlags, conn *models.Connection) error {
-	client := newAPIClient(mcCtx)
-
-	existing, _ := client.GetConnection(flags.Name, flags.Namespace)
-	if existing != nil {
-		conn.ID = existing.ID
-		conn.CreatedAt = existing.CreatedAt
-	} else {
-		conn.ID = uuid.New()
-	}
-
-	if flags.Test {
-		if err := client.SaveConnection(conn); err != nil {
-			return fmt.Errorf("failed to save connection: %w", err)
-		}
-		result, err := client.TestConnection(conn.ID.String())
-		if err != nil {
-			return fmt.Errorf("connection test failed: %w", err)
-		}
-		clicky.MustPrint(result, clicky.Flags.FormatOptions)
-		fmt.Println("\nConnection test passed")
-		return nil
-	}
-
-	if err := client.SaveConnection(conn); err != nil {
-		return fmt.Errorf("failed to save connection: %w", err)
-	}
-
-	action := "created"
-	if existing != nil {
-		action = "updated"
-	}
-	fmt.Printf("Connection '%s' %s in namespace '%s'\n", flags.Name, action, flags.Namespace)
-	return nil
-}
-
-func runConnectionAddViaDB(flags *connectionFlags, conn *models.Connection) error {
-	ctx, stop, err := duty.Start("mission-control", duty.ClientOnly)
-	if err != nil {
-		return err
-	}
-	shutdown.AddHookWithPriority("database", shutdown.PriorityCritical, stop)
-	defer stop()
-
-	var existing models.Connection
-	err = ctx.DB().Where("name = ? AND namespace = ? AND deleted_at IS NULL", flags.Name, flags.Namespace).First(&existing).Error
-	isUpdate := false
-	if err == nil {
-		isUpdate = true
-	} else if !errors.Is(err, gorm.ErrRecordNotFound) {
-		return fmt.Errorf("failed to check existing connection: %w", err)
-	}
-
-	if isUpdate {
-		conn.ID = existing.ID
-		conn.CreatedAt = existing.CreatedAt
-	} else {
-		conn.ID = uuid.New()
-	}
-
-	if flags.Test {
-		hydrated, err := ctx.HydrateConnection(conn)
-		if err != nil {
-			return fmt.Errorf("failed to hydrate connection: %w", err)
-		}
-		result, err := connection.Test(ctx, hydrated)
-		if err != nil {
-			clicky.MustPrint(result, clicky.Flags.FormatOptions)
-			return fmt.Errorf("connection test failed: %w", err)
-		}
-		clicky.MustPrint(result, clicky.Flags.FormatOptions)
-		fmt.Println("\nConnection test passed")
-	}
-
-	if err := ctx.DB().Save(conn).Error; err != nil {
-		return fmt.Errorf("failed to save connection: %w", err)
-	}
-
-	if isUpdate {
-		fmt.Printf("Connection '%s' updated in namespace '%s'\n", flags.Name, flags.Namespace)
-	} else {
-		fmt.Printf("Connection '%s' created in namespace '%s'\n", flags.Name, flags.Namespace)
-	}
-
-	return nil
-}
-
-func buildConnectionFromFlags(flags *connectionFlags) (models.Connection, error) {
+func BuildConnectionFromFlags(flags *ConnectionFlags) (models.Connection, error) {
 	conn := models.Connection{
 		Name:        flags.Name,
 		Namespace:   flags.Namespace,
@@ -691,7 +234,7 @@ func buildConnectionFromFlags(flags *connectionFlags) (models.Connection, error)
 	return conn, nil
 }
 
-func addCommonFlags(cmd *cobra.Command, flags *connectionFlags) {
+func addCommonFlags(cmd *cobra.Command, flags *ConnectionFlags) {
 	cmd.Flags().StringVar(&flags.Name, "name", "", "Connection name (required)")
 	cmd.Flags().StringVar(&flags.Namespace, "namespace", "", "Connection namespace (required)")
 	cmd.Flags().BoolVar(&flags.Test, "test", false, "Test connection before saving")
@@ -700,7 +243,7 @@ func addCommonFlags(cmd *cobra.Command, flags *connectionFlags) {
 	_ = cmd.MarkFlagRequired("namespace")
 }
 
-func addTypeSpecificFlags(cmd *cobra.Command, flags *connectionFlags, connType string) {
+func addTypeSpecificFlags(cmd *cobra.Command, flags *ConnectionFlags, connType string) {
 	switch connType {
 	case models.ConnectionTypeSlack:
 		cmd.Flags().StringVar(&flags.Channel, "channel", "", "Slack channel ID (required)")
@@ -905,33 +448,7 @@ func addTypeSpecificFlags(cmd *cobra.Command, flags *connectionFlags, connType s
 	}
 }
 
-func addConnectionFlags(cmd *cobra.Command, flags *connectionFlags, connType string) {
+func addConnectionFlags(cmd *cobra.Command, flags *ConnectionFlags, connType string) {
 	addCommonFlags(cmd, flags)
 	addTypeSpecificFlags(cmd, flags, connType)
-}
-
-func newConnectionAddTypeCommand(spec connectionTypeSpec) *cobra.Command {
-	flags := &connectionFlags{}
-	cmd := &cobra.Command{
-		Use:               spec.Name,
-		Aliases:           spec.Aliases,
-		Short:             spec.Short,
-		DisableAutoGenTag: true,
-		RunE: func(cmd *cobra.Command, args []string) error {
-			flags.Type = spec.Type
-			return runConnectionAdd(flags)
-		},
-	}
-	addConnectionFlags(cmd, flags, spec.Type)
-
-	return cmd
-}
-
-func init() {
-	for _, spec := range connectionAddTypeSpecs {
-		ConnectionAdd.AddCommand(newConnectionAddTypeCommand(spec))
-	}
-
-	Connection.AddCommand(ConnectionAdd)
-	Root.AddCommand(Connection)
 }
