+++ clientcmd/connection_test_cmd.go
@@ -1,22 +1,13 @@
-package cmd
+package clientcmd
 
 import (
-	gocontext "context"
 	"errors"
 	"fmt"
 	"os"
 
 	"github.com/flanksource/clicky"
-	"github.com/flanksource/duty"
-	"github.com/flanksource/duty/context"
-	"github.com/flanksource/duty/models"
-	"github.com/flanksource/duty/shutdown"
 	"github.com/spf13/cobra"
-	"sigs.k8s.io/yaml"
 
-	v1 "github.com/flanksource/incident-commander/api/v1"
-	"github.com/flanksource/incident-commander/connection"
-	"github.com/flanksource/incident-commander/db"
 	"github.com/flanksource/incident-commander/sdk"
 )
 
@@ -37,7 +28,6 @@ Examples:
 
   # Test an existing connection with URL override
   app connection test http --name my-http --url https://other-endpoint.com`,
-	PersistentPreRun:  PreRun,
 	SilenceUsage:      true,
 	DisableAutoGenTag: true,
 }
@@ -48,56 +38,16 @@ var (
 	connectionTestNamespace string
 )
 
-func hydrateAndTest(conn *models.Connection) (any, error) {
-	ctx := context.NewContext(gocontext.Background())
-	hydrated, err := ctx.HydrateConnection(conn)
-	if err != nil {
-		return nil, fmt.Errorf("failed to hydrate connection: %w", err)
-	}
-
-	result, err := connection.Test(ctx, hydrated)
-	if err != nil {
-		return result, fmt.Errorf("connection test failed: %w", err)
-	}
-
-	return result, nil
-}
-
-func runConnectionTestFromDB(name, namespace string, overrides *connectionFlags) (any, error) {
-	if mcCtx, ok := contextHasAPI(); ok {
+// runConnectionTestFromDB tests a saved connection. With an API context it runs
+// remotely; otherwise it delegates to the local DB implementation when present.
+func runConnectionTestFromDB(name, namespace string, overrides *ConnectionFlags) (any, error) {
+	if mcCtx, ok := ContextHasAPI(); ok {
 		return runConnectionTestViaAPI(mcCtx, name, namespace)
 	}
-
-	ctx, stop, err := duty.Start("mission-control", duty.ClientOnly)
-	if err != nil {
-		return nil, err
+	if LocalConnections == nil {
+		return nil, errNoLocalConnections
 	}
-	shutdown.AddHookWithPriority("database", shutdown.PriorityCritical, stop)
-
-	var conn models.Connection
-	if err := ctx.DB().Where("name = ? AND namespace = ? AND deleted_at IS NULL", name, namespace).First(&conn).Error; err != nil {
-		return nil, fmt.Errorf("connection %s/%s not found: %w", namespace, name, err)
-	}
-
-	if overrides != nil {
-		applyConnectionOverrides(&conn, overrides)
-	}
-
-	if clicky.Flags.LevelCount >= 1 {
-		printConnectionState(conn, clicky.Flags.LevelCount)
-	}
-
-	hydrated, err := ctx.HydrateConnection(&conn)
-	if err != nil {
-		return nil, fmt.Errorf("failed to hydrate connection: %w", err)
-	}
-
-	result, err := connection.Test(ctx, hydrated)
-	if err != nil {
-		return result, fmt.Errorf("connection test failed: %w", err)
-	}
-
-	return result, nil
+	return LocalConnections.TestSaved(name, namespace, overrides)
 }
 
 func runConnectionTestViaAPI(mcCtx *MCContext, name, namespace string) (any, error) {
@@ -106,7 +56,7 @@ func runConnectionTestViaAPI(mcCtx *MCContext, name, namespace string) (any, err
 		return result, err
 	}
 
-	upgraded, upErr := ensureAPIBase(mcCtx)
+	upgraded, upErr := EnsureAPIBase(mcCtx)
 	if upErr != nil {
 		return nil, fmt.Errorf("%w (probe failed: %v)", err, upErr)
 	}
@@ -118,7 +68,7 @@ func runConnectionTestViaAPI(mcCtx *MCContext, name, namespace string) (any, err
 }
 
 func callConnectionTestAPI(mcCtx *MCContext, name, namespace string) (any, error) {
-	client := newAPIClient(mcCtx)
+	client := NewAPIClient(mcCtx)
 
 	conn, err := client.GetConnection(name, namespace)
 	if err != nil {
@@ -132,64 +82,17 @@ func callConnectionTestAPI(mcCtx *MCContext, name, namespace string) (any, error
 	return result.Payload, nil
 }
 
-func applyConnectionOverrides(conn *models.Connection, flags *connectionFlags) {
-	if flags.URL != "" {
-		conn.URL = flags.URL
+// runConnectionTestTransient tests a connection built from flags (not saved).
+// This requires local hydration, so it is only available in the full binary.
+func runConnectionTestTransient(flags *ConnectionFlags) (any, error) {
+	if LocalConnections == nil {
+		return nil, fmt.Errorf("testing an unsaved connection %w; use --name to test a saved connection remotely", errNoLocalConnections)
 	}
-	if flags.Username != "" {
-		conn.Username = flags.Username
-	}
-	if flags.Password != "" {
-		conn.Password = flags.Password
-	}
-	if flags.Certificate != "" {
-		conn.Certificate = flags.Certificate
-	}
-	if flags.InsecureTLS {
-		conn.InsecureTLS = true
-	}
-}
-
-func runConnectionTest(flags *connectionFlags) (any, error) {
-	if flags.FromProfile != "" {
-		if err := loadAWSProfile(flags); err != nil {
-			return nil, err
-		}
-	}
-
-	conn, err := buildConnectionFromFlags(flags)
-	if err != nil {
-		return nil, fmt.Errorf("failed to build connection: %w", err)
-	}
-
-	return hydrateAndTest(&conn)
-}
-
-func runConnectionTestFromFile(filename string) (any, error) {
-	data, err := os.ReadFile(filename)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read file: %w", err)
-	}
-
-	var crd v1.Connection
-	if err := yaml.Unmarshal(data, &crd); err != nil {
-		return nil, fmt.Errorf("failed to parse YAML: %w", err)
-	}
-
-	if crd.Kind != "" && crd.Kind != "Connection" {
-		return nil, fmt.Errorf("expected Kind=Connection, got %s", crd.Kind)
-	}
-
-	conn, err := db.ConnectionFromCRD(&crd)
-	if err != nil {
-		return nil, err
-	}
-
-	return hydrateAndTest(&conn)
+	return LocalConnections.TestTransient(flags)
 }
 
 func newConnectionTestTypeCommand(spec connectionTypeSpec) *cobra.Command {
-	flags := &connectionFlags{}
+	flags := &ConnectionFlags{}
 	cmd := &cobra.Command{
 		Use:               spec.Name,
 		Aliases:           spec.Aliases,
@@ -211,7 +114,7 @@ func newConnectionTestTypeCommand(spec connectionTypeSpec) *cobra.Command {
 			}
 			flags.Name = "test"
 			flags.Namespace = "default"
-			result, err := runConnectionTest(flags)
+			result, err := runConnectionTestTransient(flags)
 			if err != nil {
 				if result != nil {
 					clicky.MustPrint(result, clicky.Flags.FormatOptions)
@@ -248,7 +151,10 @@ func init() {
 			return nil
 		}
 		if connectionTestFile != "" {
-			result, err := runConnectionTestFromFile(connectionTestFile)
+			if LocalConnections == nil {
+				return fmt.Errorf("testing from a CRD file %w", errNoLocalConnections)
+			}
+			result, err := LocalConnections.TestFile(connectionTestFile)
 			if err != nil {
 				if result != nil {
 					clicky.MustPrint(result, clicky.Flags.FormatOptions)
