+++ clientcmd/plugin_dispatch_test.go
@@ -1,4 +1,4 @@
-package cmd
+package clientcmd
 
 import (
 	"bytes"
