+++ clientcmd/plugin_help.go
@@ -1,4 +1,4 @@
-package cmd
+package clientcmd
 
 import (
 	"fmt"
