+++ clientcmd/plugin_help_test.go
@@ -1,4 +1,4 @@
-package cmd
+package clientcmd
 
 import (
 	"path/filepath"
