+++ clientcmd/suite_test.go
@@ -1,0 +1,13 @@
+package clientcmd
+
+import (
+	"testing"
+
+	ginkgo "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestClientCmd(t *testing.T) {
+	RegisterFailHandler(ginkgo.Fail)
+	ginkgo.RunSpecs(t, "ClientCmd")
+}
