+++ clientcmd/whoami_test.go
@@ -1,4 +1,4 @@
-package cmd
+package clientcmd
 
 import (
 	"context"
