+++ sdk/catalog.go
@@ -1,0 +1,82 @@
+package sdk
+
+import (
+	"context"
+	"fmt"
+	"strings"
+
+	"github.com/flanksource/duty/models"
+	"github.com/flanksource/duty/query"
+	"github.com/google/uuid"
+)
+
+// CatalogRelationships mirrors the server's catalog relationships response
+// (catalog.ConfigRelationshipsResponse) without importing the heavy catalog
+// package. Both directions are config trees rooted at ID.
+type CatalogRelationships struct {
+	ID       uuid.UUID             `json:"id"`
+	Incoming *query.ConfigTreeNode `json:"incoming"`
+	Outgoing *query.ConfigTreeNode `json:"outgoing"`
+}
+
+// SearchCatalog runs a resource search against the remote server
+// (POST /resources/search).
+func (c *Client) SearchCatalog(ctx context.Context, req query.SearchResourcesRequest) (*query.SearchResourcesResponse, error) {
+	r, err := c.R(ctx).Post(c.apiPath("/resources/search"), req)
+	if err != nil {
+		return nil, err
+	}
+	if !r.IsOK() {
+		body, _ := r.AsString()
+		if looksLikeHTML(r.Header.Get("Content-Type"), body) {
+			return nil, ErrHTMLResponse
+		}
+		return nil, fmt.Errorf("server returned %d: %s", r.StatusCode, strings.TrimSpace(body))
+	}
+	var out query.SearchResourcesResponse
+	if err := decodeJSON(r, &out); err != nil {
+		return nil, err
+	}
+	return &out, nil
+}
+
+// GetCatalogItem fetches a single catalog item by id (GET /resources/:id).
+func (c *Client) GetCatalogItem(ctx context.Context, id string) (*models.ConfigItem, error) {
+	r, err := c.R(ctx).Get(c.apiPath("/resources/" + id))
+	if err != nil {
+		return nil, err
+	}
+	if !r.IsOK() {
+		body, _ := r.AsString()
+		if looksLikeHTML(r.Header.Get("Content-Type"), body) {
+			return nil, ErrHTMLResponse
+		}
+		return nil, fmt.Errorf("server returned %d: %s", r.StatusCode, strings.TrimSpace(body))
+	}
+	var out models.ConfigItem
+	if err := decodeJSON(r, &out); err != nil {
+		return nil, err
+	}
+	return &out, nil
+}
+
+// GetCatalogRelationships fetches the incoming/outgoing config tree for a
+// catalog item (GET /catalog/:id/relationships).
+func (c *Client) GetCatalogRelationships(ctx context.Context, id string) (*CatalogRelationships, error) {
+	r, err := c.R(ctx).Get(c.apiPath("/catalog/" + id + "/relationships"))
+	if err != nil {
+		return nil, err
+	}
+	if !r.IsOK() {
+		body, _ := r.AsString()
+		if looksLikeHTML(r.Header.Get("Content-Type"), body) {
+			return nil, ErrHTMLResponse
+		}
+		return nil, fmt.Errorf("server returned %d: %s", r.StatusCode, strings.TrimSpace(body))
+	}
+	var out CatalogRelationships
+	if err := decodeJSON(r, &out); err != nil {
+		return nil, err
+	}
+	return &out, nil
+}
